+++ tests/bootstrap.php
@@ -20,6 +20,7 @@ use Cake\Log\Log;
 use Cake\Mailer\Email;
 use Cake\Mailer\TransportFactory;
 use Cake\TestSuite\TestEmailTransport;
+use DatabaseBackup\Utility\BackupManager;
 
 date_default_timezone_set('UTC');
 mb_internal_encoding('UTF-8');
@@ -111,4 +112,4 @@ Configure::write('DatabaseBackup.mailSender', 'sender@example.com');
 Configure::write('pluginsToLoad', ['DatabaseBackup']);
 
 require_once ROOT . 'config' . DS . 'bootstrap.php';
-echo 'Running tests for `' . DATABASE_BACKUP_DRIVER . '` driver ' . PHP_EOL;
+echo 'Running tests for `' . BackupManager::getDriverName() . '` driver ' . PHP_EOL;